Fix bad merge
